[Bug #158909217] Switch custom import_assets command
