+++ core/management/commands/__init__.py
@@ -1,0 +1,7 @@
+import django
+
+__version__ = '1.0.0'
+
+
+COMMAND_VERSION = __version__
+DJANGO_VERSION = django.get_version()
